# use include_str! to trigger a rebuild when the input spec is modified